extract buoys
