+++ Slider/Assets/Scripts/Map/Ocean/FoggySeasManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ee12cc20d515f4a4cadc90cd4689c2f9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
