+++ Slider/Assets/Scripts/Map/Ocean/OceanBuoys.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a2cc5e78873ee984c908742b4ebabd50
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
